refactor: arborescence src/templates et BaseController (helpers + docblocks)
