+++ src/Controller/Front/HomeController.php
@@ -1,17 +1,17 @@
 <?php
 
-namespace App\Controller;
+namespace App\Controller\Front;
 
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use App\Controller\Base\BaseController;
 
-final class HomeController extends AbstractController
+final class HomeController extends BaseController
 {
     #[Route('/', name: 'app_home')]
     public function index(): Response
     {
-        return $this->render('home/index.html.twig', [
+        return $this->render('front/home/index.html.twig', [
             'controller_name' => 'HomeController',
         ]);
     }
